Revert "Fix branch auth"
