+++ package.json
@@ -12,7 +12,6 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.0",
     "@angular/animations": "^4.0.0",
     "@angular/common": "^4.0.0",
     "@angular/compiler": "^4.0.0",
@@ -22,23 +21,10 @@
     "@angular/platform-browser": "^4.0.0",
     "@angular/platform-browser-dynamic": "^4.0.0",
     "@angular/router": "^4.0.0",
-    "angular2-image-upload": "^0.6.1",
     "angular2-token": "^0.2.0-beta.11",
-    "bootstrap": "^3.3.7",
     "core-js": "^2.4.1",
-    "ngx-uploader": "^3.0.6",
-    "angular2-feature-toggle": "^1.0.3",
-    "angular2-google-maps": "^0.17.0",
-    "angular2-slimscroll": "^0.3.0",
-    "angular2-token": "^0.2.0-beta.11",
-    "auth0-js": "^8.7.0",
-    "auth0-lock-passwordless": "^2.2.3",
-    "bootstrap": "^3.3.7",
-    "core-js": "^2.4.1",
-    "ng2-dropdown": "0.0.21",
     "ng2-emoji": "^0.2.0",
-    "ngx-uploader": "^3.0.6",
-    "ng2-filter-pipe": "^0.1.10",
+    "angular2-emoji": "^0.1.6",
     "rxjs": "^5.1.0",
     "zone.js": "^0.8.4"
   },
@@ -46,19 +32,17 @@
     "@angular/cli": "1.1.1",
     "@angular/compiler-cli": "^4.0.0",
     "@angular/language-service": "^4.0.0",
-    "@types/googlemaps": "^3.26.13",
     "@types/jasmine": "2.5.45",
     "@types/node": "~6.0.60",
-    "angular2-perfect-scrollbar": "^2.0.6",
     "codelyzer": "~3.0.1",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
