+++ src/app/app.component.html
@@ -1,1 +1,2 @@
-<router-outlet></router-outlet>
+<app-user></app-user>
+<app-conversation></app-conversation>
