--- src/app/app.constants.ts
@@ -1,3 +1,0 @@
-export class AppSettings {
-   public static API_ENDPOINT='http://localhost:3000/';
-}
