--- src/app/auth/auth.component.html
@@ -1,12 +1,0 @@
-<div class="container col-sm-4 form-group">
-  <h1>Welcome to DihlApp!</h1>
-
-  <form (submit)="sendMail(email)" #form="ngForm">
-    <fieldset>
-      <input type="email" class="form-control" name="email" id="email" [(ngModel)]="email" required />
-      <label for="email" [ngClass]="{'stay': email.length > 0}">Email</label>
-    </fieldset>
-
-    <button type="submit" class="btn btn-primary" [disabled]="!form.valid">Send email</button>
-  </form>
-</div>
