--- src/app/auth/auth.component.scss
@@ -1,33 +1,0 @@
-.container {
-  fieldset {
-    border: none;
-    position: relative;
-    input {
-      height: 60px;
-      margin: 15px;
-      &:focus + label {
-        font-weight: bold;
-        font-size: .7em;
-        top: 2em;
-      }
-    }
-    label {
-      cursor: text;
-      position: absolute;
-      top: 35%;
-      left: 6%;
-      color: #333;
-      font-weight: normal;
-      font-family: sans-serif, Arial;
-      transition: .15s ease-in all;
-    }
-    .stay {
-      font-weight: bold;
-      font-size: .7em;
-      top: 2em;
-    }
-  }
-  .btn {
-    margin-left: 15px;
-  }
-}
