+++ src/app/conversation/conversation.component.scss
@@ -1,0 +1,57 @@
+.chatbox {
+  float: left;
+  background: #ecf2f5;
+  width: 50%;
+  height: 100%;
+  overflow-y: scroll;
+  padding: 20px;
+
+  :host >>> .emogi-image {
+    height: 60px;
+    width: 60px;
+    border-radius: 50%;
+  }
+  
+  .user {
+    width: 100%;
+    margin-top: 10px;
+    float: left;
+    img {
+      float: left;
+      border-radius: 50%;
+      width: 50px;
+      margin-right: 5px;
+    }
+
+    p {
+      float: left;
+      margin: 0 10px;
+      line-height: 50px;
+    }
+
+    i {
+      height: 16px;
+    }
+  }
+  .current-user {
+    width: 100%;
+    margin-top: 10px;
+    float: right;
+    img {
+      float: right;
+      border-radius: 50%;
+      width: 50px;
+      margin-left: 5px;
+    }
+
+    p {
+      float: right;
+      margin: 0 10px;
+      line-height: 50px;
+    }
+
+    i {
+      height: 16px;
+    }
+  }
+}
