+++ src/app/conversation/message.type.ts
@@ -1,0 +1,13 @@
+export class Message {
+  id: number;
+  content: string;
+  sent: boolean;
+  seen: boolean;
+  timestamp: string;
+  user: User;
+}
+
+export class User {
+  name: string;
+  avatar: string;
+}
