--- src/app/main-view/conversation/conversation.component.html
@@ -1,51 +1,0 @@
-<section class="chatbox">
-  <!-- <div class="typing-status">
-    <p *ngIf="contact.isTyping" class="light"><span class="bold">{{contact.name}}</span>is typing ...</p>
-    <button class="fa fa-star" aria-hidden="true" (click)="Favorite(contact)" [ngClass]="{yellow: isFavorite}"></button>
-  </div> -->
-  <div class="conversation clearfix">
-    <div class="message" *ngFor="let message of messages" [ngClass]="message.user.name === 'Joe' ? 'current-user' : 'user'">
-      <img src="{{message.user.avatar}}"/>
-      <p>{{message.content}}</p>
-      <!-- <button (click)="deleteMessage(message.id)">Delete</button> -->
-      <p class="light">{{message.timestamp}}</p>
-      <p class="seen" *ngIf="message.seen">
-        <span class="fa-stack">
-          <i class="fa fa-check fa-stack-1x" style="margin-left:4px"></i>
-          <i class="fa fa-check  fa-stack-1x" style="margin-left:-4px"></i>
-        </span>
-      </p>
-      <p class="sent" *ngIf="message.sent && message.seen === false">
-        <i class="fa fa-check"></i>
-      </p>
-    </div>
-  </div>
-
-  <form class="send-message" (submit)="sendMessage(message)" >
-      <emoji-input>
-        [onEnter]="onEnterFunction"
-        [popupAnchor]="'bottom'"
-        [(model)]="message.content">
-      </emoji-input>
-      <div class="attachment-menu" [hidden]="attachmentMenu">
-        <ul>
-          <li>
-            <app-location></app-location>
-          </li>
-          <li>
-            <button type="button" (click)="fakeClick()" class="uploadButton"><i class="fa fa-upload" aria-hidden="true"></i></button>
-            <input type="file" (change)="changeListner($event)" class="hiddenUploadButton" id="fileSelector">
-            <!-- <img src="{{tempImg}}" /> -->
-          </li>
-        </ul>
-      </div>
-      <button class="att-button" name="attachment" type="button" (click)="showAttachmentMenu()"><i class="fa fa-paperclip" aria-hidden="true"></i></button>
-      <button type="submit" class="submit-message"><i class="fa fa-paper-plane" aria-hidden="true"></i></button>
-    <!-- <button type="button" (click)="fakeClick()" class="uploadButton">Choose a file</button>
-    <input type="file" (change)="changeListner($event)" class="hiddenUploadButton" id="fileSelector"> -->
-    <!-- <textarea placeholder="Type your message" [(ngModel)]="message.content"></textarea> -->
-
-    <!-- <button type="submit" (click)="startUpload()" class="send-message"><i class="fa fa-paper-plane" aria-hidden="true"></i></button> -->
-
-  </form>
-</section>
