--- src/app/main-view/conversation/conversation.component.scss
@@ -1,178 +1,0 @@
-.chatbox {
-  float: left;
-  background: #ecf2f5;
-  width: 50%;
-  height: 100%;
-  overflow-y: scroll;
-  padding: 20px;
-
-  &::-webkit-scrollbar {
-    width: 6px;
-    background-color: #F5F5F5;
-    border-radius: 10px;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    background-color: #ccc;
-    border-radius: 10px;
-  }
-
-  &::-webkit-scrollbar-track {
-    -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
-    background-color: #F5F5F5;
-  }
-
-  form.send-message {
-    position: absolute;
-    bottom: 0;
-    width: 50%;
-    emoji-input {
-      position: relative;
-      margin-left: -20px;
-      width: 90%;
-      float: left;
-    }
-    .att-button,
-    .submit-message{
-      position: relative;
-      display: inline-block;
-      float: left;
-      background-color: #fff;
-      height: 37px;
-      border-style: ridge;
-      border: 1px solid #ccc;
-      color: #e2e2e2;
-    }
-    .submit-message{
-      color: #777;
-    }
-    .attachment-menu{
-      position: absolute;
-      bottom: 110%;
-      right: 10%;
-      li{
-        list-style: none;
-        float: left;
-      }
-    }
-
-  }
-  /*  .conversation {
-    .message {
-      width: 50%;
-    }
-
-    .bubble{
-      background-color: #F2F2F2;
-      max-width: 50%;
-      border-radius: 5px;
-      box-shadow: 0 0 6px #B2B2B2;
-      display: inline-block;
-      padding: 10px 18px;
-      position: relative;
-      vertical-align: top;
-      &::before {
-        background-color: #F2F2F2;
-        content: "\00a0";
-        display: block;
-        height: 16px;
-        position: absolute;
-        top: 11px;
-        transform:             rotate( 29deg ) skew( -35deg );
-        -moz-transform:    rotate( 29deg ) skew( -35deg );
-        -ms-transform:     rotate( 29deg ) skew( -35deg );
-        -o-transform:      rotate( 29deg ) skew( -35deg );
-        -webkit-transform: rotate( 29deg ) skew( -35deg );
-        width:  20px;
-      }
-    }
-
-    .you {
-      float: left;
-      margin: 5px 45px 5px 5px;
-      &::before {
-        box-shadow: -2px 2px 2px 0 rgba( 178, 178, 178, .4 );
-        left: -9px;
-      }
-    }
-
-    .me {
-      float: right;
-      margin: 5px 5px 5px 45px;
-      &::before {
-        box-shadow: 2px -2px 2px 0 rgba( 178, 178, 178, .4 );
-        right: -9px;
-      }
-    }
-  }*/
-  :host >  >  > .emogi-image {
-    height: 60px;
-    width: 60px;
-    border-radius: 50%;
-    /*  .conversation{
-     min-height: calc(100vh - 70px);
-    :host >>> .emogi-image {
-      height: 60px;
-      width: 60px;
-      border-radius: 50%;
-    }
-    .send-message{
-      height: 100px;
-      margin:10px auto;
-    }*/
-  }
-
-  div.emoji-search.top {
-    bottom: 36px;
-  }
-
-  .user {
-    width: 100%;
-    margin-top: 10px;
-    float: left;
-
-    img {
-      float: left;
-      border-radius: 50%;
-      width: 50px;
-      margin-right: 5px;
-    }
-
-    p {
-      float: left;
-      margin: 0 10px;
-      line-height: 50px;
-    }
-
-    i {
-      height: 16px;
-    }
-  }
-
-  .current-user {
-    width: 100%;
-    margin-top: 10px;
-    float: right;
-
-    img {
-      float: right;
-      border-radius: 50%;
-      width: 50px;
-      margin-left: 5px;
-    }
-
-    p {
-      float: right;
-      margin: 0 10px;
-      line-height: 50px;
-    }
-
-    i {
-      height: 16px;
-    }
-  }
-}
-
-.hiddenUploadButton {
-  display: none;
-}
