--- src/app/main-view/conversation/location/location.component.html
@@ -1,19 +1,0 @@
-
-<!-- <button type="button" (click)="hideMap=false" name="button">show map</button> -->
-<button name="location" (click)="hideMap=false; doTriggerResize()" type="button"><i class="fa fa-map-marker" aria-hidden="true"></i></button>
-<div class="overlay" [ngClass]="{'hide-map': hideMap}">
-  <div class="map-modal">
-    <div class="form-group">
-      <input placeholder="search for location" autocorrect="off" autocapitalize="off" spellcheck="off" type="text" class="form-control" #search >
-    </div>
-    <div class="map-container">
-      <agm-map #myMap class="map" [latitude]="latitude" [longitude]="longitude" [scrollwheel]="false" [zoom]="zoom">
-        <agm-marker [markerDraggable]="true" [latitude]="latitude" [longitude]="longitude" (dragEnd)="updateMarkerPos($event)"></agm-marker>
-      </agm-map>
-    </div>
-    <button class="send-location" type="button" (click)="staticMapUrl(15,310)">Send Location</button>
-    <button class="cancel-location" type="button" (click)="hideMap=true">Cancel</button>
-  </div>
-</div>
-
-<!-- <img src="{{staticMap}}" alt=""> -->
