--- src/app/main-view/conversation/location/location.component.scss
@@ -1,64 +1,0 @@
-.hide-map{
-  display: none;
-  position: fixed;
-  z-index: -1;
-}
-.background-close{
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: -1;
-  opacity: 0;
-}
-
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: 2;
-  background-color: rgba(17, 17, 17, 0.64);
-
-
-  .map-modal {
-    position: relative;
-    height: 500px;
-    width: 500px;
-    top: 50%;
-    left: 50%;
-    z-index: 3;
-    background-color: #DDD;
-    margin-top: -250px;
-    margin-left: -250px;
-    border: 1px solid #777;
-    border-radius: 15px;
-
-    .map-container {
-      position: relative;
-      height: 400px;
-      width: 400px;
-      margin: 0 auto;
-      .map {
-        height: 100%;
-        width: 100%;
-      }
-    }
-    button{
-      position: absolute;
-      border-radius: 10px;
-      background-color: #851123;
-      color: #CCC;
-      margin-top: 10px;
-
-    }
-    .send-location {
-      left: 75%;
-    }
-    .cancel-location{
-      left: 60%;
-    }
-  }
-}
