--- src/app/main-view/conversation/message.type.ts
@@ -1,27 +1,0 @@
-export interface Message {
-  id: number;
-  content: string;
-  sent: boolean;
-  seen: boolean;
-  timestamp: string;
-  user: User;
-}
-
-interface User {
-  name: string;
-  avatar: string;
-}
-
-// export class Message {
-//   id: number;
-//   content: string;
-//   sent: boolean;
-//   seen: boolean;
-//   timestamp: string;
-//   user: User;
-// }
-
-// export class User {
-//   name: string;
-//   avatar: string;
-// }
