--- src/app/main-view/main-view.component.html
@@ -1,4 +1,0 @@
-<app-user></app-user>
-<app-conversation></app-conversation>
-<app-rightbar></app-rightbar>
-
