--- src/app/main-view/main-view.component.scss
@@ -1,1 +1,0 @@
-
