--- src/app/main-view/rightbar/rightbar.component.html
@@ -1,20 +1,0 @@
-<section class="rightside">
-  <div class="user">
-  <h1>Welcome to Dihlapp</h1>
-  <!-- <h1>Welcome to Dihlapp, {{currentUser.name}}</h1> -->
-  <img src="https://placeholdit.co//i/200x200" class="avatar"/>
-  <p>This is a dummy status message</p>
-  <!-- <p>This is a dummy status message{{currentUser.personalMessage}}</p> -->
-</div>
-<div class="features">
-  <h1>Features</h1>
-  <ul class="feature clearfix">
-        <li><a href="#"><i class="fa fa-google fa-4x" aria-hidden="true"></i></a></li>
-        <li><a href="#"><i class="fa fa-youtube fa-4x" aria-hidden="true"></i></a></li>
-        <li><a href="#"><i class="fa fa-soundcloud fa-4x" aria-hidden="true"></i></a></li>
-        <li><a href="#"><i class="fa fa-map-marker fa-4x" aria-hidden="true"></i></a></li>
-        <li><a href="#"><i class="fa fa-calendar fa-4x" aria-hidden="true"></i></a></li>
-        <li><a href="#"><i class="fa fa-plus fa-4x" aria-hidden="true"></i></a></li>
-  </ul>
-</div>
-</section>
