--- src/app/main-view/rightbar/rightbar.component.scss
@@ -1,72 +1,0 @@
-.rightside{
-  width:25%;
-  height:100%;
-  background: #485361;
-  text-align: center;
-  display:inline-block;
-  padding:15px 0;
-  color:white;
-  overflow-y: scroll;
-  overflow-x: hidden;
-  &::-webkit-scrollbar {
-    width: 6px;
-    background-color: #F5F5F5;
-    border-radius: 10px;
-  }
-  &::-webkit-scrollbar-thumb {
-    background-color: #333;
-    border-radius: 10px;
-  }
-  &::-webkit-scrollbar-track {
-    -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
-    background-color: #F5F5F5;
-    border-radius: 10px;
-  }
-  .features{
-    border-top: 1px solid rgba(145, 150, 156, 0.25);
-    margin: 20px 0;
-  }
-  img{
-    border-radius: 50%;
-    margin:15px 0;
-  }
-    h1, p{
-      padding:0;
-      display:block;
-      margin:25px 20px;
-    }
-    p{
-      font-size: 20px;
-    }
-.features{
-  display: inline-block;
-  margin: 0 auto;
-  margin:0 5px;
-
-  // padding:20px 5px;
-  ul{
-    list-style: none;
-    padding:0;
-    position: relative;
-
-
-    li{
-      width:10vw;
-      height:10vw;
-      background:#2b3035;
-      border-radius:50%;
-      padding:15px;
-      display: inline-block;
-      margin: 0 auto;
-      margin:10px 5px;
-      text-align: center;
-      a{
-        color:white;
-        i{
-          font-size: 5vw;
-        }
-      }
-    }
-  }
-}
-}
