--- src/app/main-view/rightbar/rightbar.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-rightbar',
-  templateUrl: './rightbar.component.html',
-  styleUrls: ['./rightbar.component.scss']
-})
-export class RightbarComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
