--- src/app/main-view/user/user.component.scss
@@ -1,183 +1,0 @@
-.contacts {
-  padding: 10px;
-  float: left;
-  background: #485361;
-  width: 25%;
-  height: 100%;
-  overflow-y: scroll;
-  box-shadow: 2px 0px 8px 0px rgba(50, 50, 50, 0.9);
-  
-  &::-webkit-scrollbar {
-    width: 6px;
-    background-color: #F5F5F5;
-    border-radius: 10px;
-  }
-  &::-webkit-scrollbar-thumb {
-    background-color: #333;
-    border-radius: 10px;
-  }
-  &::-webkit-scrollbar-track {
-    -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
-    background-color: #F5F5F5;
-    border-radius: 10px;
-  }
-
-
-  /* search bar */
-
-	.search {
-		width: 90%;
-		margin: 20px auto;
-		border-radius: 20px;
-		background: #323f4e;
-		border: none;
-		padding: 0 15px;
-
-		input {
-			width: 80%;
-			font-size: 14px;
-			padding: 10px 5px;
-			background: none;
-			border: none;
-			outline: none;
-			color: #868b8e;
-			&::placeholder {
-				color: #868b8e;
-			}
-		}
-
-		.fa-search {
-			color: #fff;
-      cursor: pointer;
-		}
-	}
-
-  /* search results */
-
-  .search-results {
-    border-radius: 10px;
-    padding: 5px 20px;
-    margin-left: 15px;
-    margin-top: -10px;
-    width: 15%;
-    height: 60%;
-    position: absolute;
-    background: #eee;
-    overflow-y: scroll;
-    &::-webkit-scrollbar {
-    width: 6px;
-    background-color: #F5F5F5;
-    border-radius: 10px;
-    }
-    &::-webkit-scrollbar-thumb {
-      background-color: #333;
-      border-radius: 10px;
-    }
-    &::-webkit-scrollbar-track {
-      -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
-      background-color: #F5F5F5;
-      border-radius: 10px;
-    }
-
-    ul {
-      list-style: none;
-      padding: 0;
-      li {
-        width: 100%;
-      }
-    }
-
-    a {
-      display: block;
-      text-align: center;
-      cursor: pointer;
-      text-decoration: none;
-      color: #333;
-      & + a {
-        border-top: 1px solid #ccc;
-        margin-top: 10px;
-        padding-top: 10px;
-      }
-    }
-  }
-
-  /* contacts list */
-
-	.contacts-list {
-		padding: 15px;
-		color: #fff;
-    border-top: 1px solid rgba(145, 150, 156, 0.25);
-    width: 100%;
-
-    .contact {
-      cursor: pointer;
-      &:hover .fa-caret-down {
-        display: block;
-      }
-    }
-
-    /* conversation menu */
-
-    .conversation-menu {
-      border-radius: 10px;
-      padding: 5px 20px;
-      left: 50%;
-      margin-top: 20px;
-      width: 15%;
-      position: absolute;
-      background: #ccc;
-      button {
-        border: none;
-        border-radius: 10px;
-        background: none;
-        outline: none;
-        color: #333;
-        font-family: Helvetica;
-        width: 100%;
-        margin: 0 auto;
-        & + button {
-          margin-top: 5px;
-        }
-        &:hover {
-          background-color: #dcdfe5;
-        }
-      }
-    }
-
-    .fa-caret-down {
-      display: none;
-      font-size: 18px;
-      float: right;
-    }
-
-		img {
-			border-radius: 50%;
-			width: 50px;
-			float: left;
-			margin-right: 20px;
-		}
-
-
-		.contact-name {
-			font-size: 16px;
-      font-family: Helvetica;
-      line-height: 1.5;
-      margin-bottom: -2px;
-		}
-
-		.contact-message,
-		.time-created {
-			color: #91969c;
-			font-size: 14px;
-      font-family: Helvetica;
-			margin-left: 70px;
-			overflow: hidden;
-			text-overflow: ellipsis;
-			white-space: nowrap;
-		}
-		.time-created {
-			float: right;
-			font-size: 10px;
-		}
-	}
-}
