--- src/app/profileupdate/profileupdate.component.html
@@ -1,39 +1,0 @@
-<body>
-  <div class="redirectForm">
-    <h1>Welcome to Dihlapp!</h1>
-    <p>Please take a moment to set up your user profile :)</p>
-    <form class="profile-update" >
-    <table>
-      <tr>
-        <label class="upload-button" for="avatar">Profile Photo</label>
-      </tr>
-      <tr>
-        <div class="avatar" id="list"><img src={{imagePath}}/></div>
-      </tr>
-      <tr>
-        <button type="button" name="button"  (click)="fakeClick()"> Upload Image</button>
-        <input type="file" id="fileSelector" style="display:none" name="image_upload" (change)="imageSelected($event)">
-      </tr>
-      <tr>
-        <label for="firstName">First Name</label>
-      </tr>
-      <tr>
-        <input type="text" name="firstName" id="firstName"><br/>
-      </tr>
-      <tr>
-        <label for="lastName">Last Name</label>
-      </tr>
-      <tr>
-        <input type="text" name="lastName" id="lastName"><br/>
-      </tr>
-      <tr>
-        <label for="personalMessage">Personal Message</label>
-      </tr>
-      <tr>
-        <input type="text" name="personalMessage" id="personalMessage"><br/>
-      </tr>
-      <button (click)="EditProfile()">Save</button>
-    </table>
-  </form>
-</div>
-</body>
