--- src/app/register/register.component.html
@@ -1,15 +1,0 @@
-<div class="container col-sm-4 registration-form">
-  <form class="form-group" #form="ngForm" (submit)="addUser()">
-    <fieldset>
-      <input type="text" class="form-control form-control-lg" name="first-name" id="first-name" [(ngModel)]="registerData.firstName" required />
-      <label for="first-name" [ngClass]="{'stay': registerData.firstName.length > 0}">First name</label>
-    </fieldset>
-
-    <fieldset>
-      <input type="text" class="form-control form-control-lg" name="last-name" id="last-name" [(ngModel)]="registerData.lastName" required />
-      <label for="last-name" [ngClass]="{'stay': registerData.lastName.length > 0}">Last name</label>
-    </fieldset>
-
-    <button type="submit" class="btn btn-primary" [disabled]="!form.valid">Submit</button>
-  </form>
-</div>
