--- src/app/register/register.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { RegisterService } from './register.service';
-
-describe('RegisterService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [RegisterService]
-    });
-  });
-
-  it('should be created', inject([RegisterService], (service: RegisterService) => {
-    expect(service).toBeTruthy();
-  }));
-});
