+++ src/app/services/message.service.ts
@@ -1,29 +1,28 @@
 import { Injectable } from '@angular/core';
 import { Http } from '@angular/http';
 
-import { Message } from '../main-view/conversation/message.type';
-import { MESSAGES } from '../main-view/conversation/mock-message';
-import { AppSettings } from '../app.constants'
+import { Message } from '../conversation/message.type';
+import { MESSAGES } from '../conversation/mock-message';
 
 import 'rxjs/add/operator/map';
 
-@Injectable()
+@Injectable()  
 export class MessageService {
 
   constructor(private http: Http) { }
 
-  getMessages(id) {
-    return this.http.get(AppSettings.API_ENDPOINT + 'conversation/'+id)
+  getMessages() {
+    return this.http.get('http://jsonplaceholder.typicode.com/posts')
                     .map(res => res.json().data)
-  }
+  }  
 
   postMessage(message) {
-    return this.http.post(AppSettings.API_ENDPOINT + 'messages', message)
+    return this.http.post('http://jsonplaceholder.typicode.com/posts', message)
                     .map(res => res.json().data)
   }
 
-  // destroyMessage(id) {
-  //   return this.http.delete('http://jsonplaceholder.typicode.com/posts/:id')
-  //                   .map(res => res.json().data)
-  // }
+  destroyMessage(id) {
+    return this.http.delete('http://jsonplaceholder.typicode.com/posts/:id')
+                    .map(res => res.json().data)
+  }
 }
