+++ src/index.html
@@ -3,11 +3,6 @@
 <head>
   <meta charset="utf-8">
   <title>Dihlappbgad</title>
-  <script src="../node_modules/angular2/bundles/angular2-polyfills.js"></script>
-  <script src="../node_modules/systemjs/dist/system.src.js"></script>
-  <script src="../node_modules/rxjs/bundles/Rx.js"></script>
-  <script src="../node_modules/angular2/bundles/angular2.dev.js"></script>
-  <script src="../node_modules/angular2/bundles/http.dev.js"></script>
   <base href="/">
 
   <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -16,7 +11,6 @@
   <link rel="stylesheet" type="text/css" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
 </head>
 <body>
-  <base href="/">
   <app-root>Loading...</app-root>
 </body>
 </html>
