+++ src/styles.scss
@@ -9,7 +9,6 @@ html {
     height: 100%;
     margin: 0;
     font-family: 'Roboto', sans-serif;
-    padding:0;
   }
 }
